% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_plots.R
\name{month_plots}
\alias{month_plots}
\title{Plot fixed-size chunks from time series data}
\usage{
month_plots(
  file,
  column = 2,
  output_dir = "../plots",
  npoints = 28 * 24 * 2,
  width_cm = 15,
  height_cm = 10,
  res = 100
)
}
\arguments{
\item{file}{Path to the CSV file to read.}

\item{column}{Column index or name of the variable to plot (default: 2).}

\item{output_dir}{Directory where PNG plots will be saved
(default: \code{"../plots"}). Created if it does not exist.}

\item{npoints}{Number of data points per plot
(default: \code{28 * 24 * 2} for one month at 30-min intervals).}

\item{width_cm}{Width of the output PNG in centimeters (default: 15).}

\item{height_cm}{Height of the output PNG in centimeters (default: 10).}

\item{res}{Resolution of output PNG in dots per inch (default: 100).}
}
\description{
This function reads a CSV file containing time series data, where the first
column must be \code{dateutc} in the format \code{"yyyy/mm/dd hh:mm:ss"}.
It splits the data into consecutive chunks of \code{npoints} rows and
generates a PNG plot for each chunk. The x-axis always spans the same number
of points, so the last chunk will have empty space if there are not enough
rows to fill it.
}
\details{
The PNG files are saved in \code{output_dir} with names formed from the base
name of the input file plus \code{"_chunkXX.png"}.
}
\examples{
\dontrun{
plot_monthly_chunks("data.csv", column = 2)
plot_monthly_chunks("data.csv", column = "tem", npoints = 1000)
}
}
