% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_filename.R
\name{parse_filename}
\alias{parse_filename}
\title{parse_filename.R
Creates an homogeneous file name following the pattern:
#' observatory.stationNumber.depth.variables.dateStart.dateEnd.csv}
\usage{
parse_filename(
  fileName,
  observatory = "mzt",
  stationNumber = 5,
  depth = "1m",
  variables = "t"
)
}
\arguments{
\item{fileName}{A character string specifying the CSV file name to parse. The file's first column must be a DateTime stamp in "YYYY/MM/DD HH:MM:SS" format.}

\item{observatory}{A string identifier for the observatory (e.g., "mzt", "cme").}

\item{stationNumber}{Station number (e.g., 1 to 11 for Mazatlan).}

\item{depth}{Measurement depth as string (e.g., "1m").}

\item{variables}{String representing measured variables (e.g., "ot" for oxygen and temperature).}
}
\value{
A character string with the standardized file name including date start and end.
}
\description{
parse_filename.R
Creates an homogeneous file name following the pattern:
#' observatory.stationNumber.depth.variables.dateStart.dateEnd.csv
}
\examples{
\dontrun{
parse_filename("mzt.boya24.20220202.csv", stationNumber = 5, variables = "ot")
}
}
