% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sea_level.R
\name{sea_level}
\alias{sea_level}
\title{Calculate sea level from submerged pressure and atmospheric pressure}
\usage{
sea_level(water_df, atm_df, salinity = 35)
}
\arguments{
\item{water_df}{A (water) data frame with columns:
\itemize{
\item \code{dateutc} (POSIXct): Timestamp of the measurement
\item \code{pres} (numeric): Absolute pressure measured by the submerged sensor (kPa)
\item \code{tem} (numeric): Water temperature in degrees Celsius
}}

\item{atm_df}{An (atmospheric) data frame with columns:
\itemize{
\item \code{dateutc} (POSIXct): Timestamp of the atmospheric measurement
\item \code{pres.y} (numeric): Atmospheric pressure in kPa
}}

\item{salinity}{Salinity in PSU (default = 35).}
}
\value{
A data frame identical to \code{water_df} but with an additional
column \code{sl} giving the estimated water column height in meters.
}
\description{
This function calculates sea level (water column height) from submerged
absolute pressure measurements and atmospheric pressure measurements.
If atmospheric pressure is missing for a given timestamp, the mean
atmospheric pressure from the atmospheric data is used instead.
}
\details{
Sea level is calculated by subtracting the atmospheric pressure from
the submerged sensor's absolute pressure, converting the resulting
gauge pressure to meters of water using the water density at the given
salinity and temperature.
}
\examples{
# water_df <- data.frame(dateutc = as.POSIXct(...), pre = ..., tem = ...)
# atm_df <- data.frame(dateutc = as.POSIXct(...), pres.y = ...)
# result <- calc_sea_level(water_df, atm_df)

}
