% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_extremes.R
\name{clean_extremes}
\alias{clean_extremes}
\title{Cleans bad data from start and end of each sonde (e.g., Hobo) record.}
\usage{
clean_extremes(data, column = "tem", n_check = 48, factor = 3)
}
\arguments{
\item{data}{A data frame containing a time series column (e.g., temperature).}

\item{column}{Column name or number to inspect (default = "tem").}

\item{n_check}{Number of rows to check at the beginning and end (default = 48).}

\item{factor}{Multiplier of standard deviation to define threshold (default = 3).}
}
\value{
A modified version of the input data frame with records with initial/last extreme values deleted.
}
\description{
This function removes extreme values from a selected numeric column in a data frame.
The criterion is based on the mean ± factor × standard deviation of the first and last
\code{n_check} values in the column. Records outside this range are assumed to be wrong and are deleted.
}
\examples{
 \dontrun{
clean_extremes(data, column = "tem", n_check = 48, factor = 3)
}
}
