% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_filename.R
\name{parse_filename}
\alias{parse_filename}
\title{parse_filename.R
jasc20250806
Parse Observatory (e.g., Hobo) File Names}
\usage{
parse_filename(
  fileName,
  observatory = "mzt",
  stationNumber = 5,
  depth = "1m",
  variables = "t"
)
}
\arguments{
\item{fileName}{A character string specifying the file name to parse. The file fist column must be a DateTime stamp.
Function and/or result parameters are:
\itemize{
\item \code{fileName}: A hobo-derived (.csv) file name ID (e.g., "mzt.5.boya24.20220202.csv").
\item \code{observatory}: a string identifier (e.g., "mzt", "cme", "pmo", "samo").
\item \code{stationNumber}: station number (e.g., 1 to 11 for Mazatlan).
\item \code{depth}: A string indicating the measurement depth (e.g., "1m").
\item \code{variables}: A string representing measured variables (e.g., "ot" for oxygen and temperature).
\item \code{dateStart}: start date (e.g., "20200101").
\item \code{dateEnd}: end date (e.g., "20200202").
\item \code{newName}: string, example: "mzt.1.2m.ot.20200101.20200202.csv"
}}
}
\value{
newName (string).
}
}
\description{
Extracts components from file names following the pattern
observatory.stationNumber.depth.variables.dateStart.dateEnd.csv
}
